+++ tests/unit/modin/test_series_strings.py
@@ -46,8 +46,6 @@ def test_str_cat_no_others(mock_str_register, mock_series):
         (lambda s: s.str.match("ab", case=False), "match"),
         (lambda s: s.str.extract("(ab)", expand=False), "extract"),
         (lambda s: s.str.extractall("(ab)", flags=1), "extractall"),
-        (lambda s: s.str.rstrip(","), "rstrip"),
-        (lambda s: s.str.lstrip(","), "lstrip"),
         (lambda s: s.str.partition("|", expand=False), "partition"),
         (lambda s: s.str.removeprefix("t"), "removeprefix"),
         (lambda s: s.str.removesuffix("a"), "removesuffix"),