+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Added support for `Series.str.get`.
 - Added support for `Series.dt.dayofweek`, `Series.dt.day_of_week`, `Series.dt.dayofyear`, and `Series.dt.day_of_year`.
 - Added support for `Series.str.__getitem__` (`Series.str[...]`).
+- Added support for `Series.str.lstrip` and `Series.str.rstrip`.
 
 #### Bug Fixes